+++ src/GRA.Controllers/ViewModel/MissionControl/Challenges/ChallengeGroupDetailViewModel.cs
@@ -1,6 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
+using GRA.Domain.Model;
 
 namespace GRA.Controllers.ViewModel.MissionControl.Challenges
 {
@@ -9,5 +8,7 @@ namespace GRA.Controllers.ViewModel.MissionControl.Challenges
         public GRA.Domain.Model.ChallengeGroup ChallengeGroup { get; set; }
         public string ChallengeIds { get; set; }
         public string Action { get; set; }
+        public List<Event> RelatedEvents { get; set; }
+        public bool CanManageEvents { get; set; }
     }
 }