+++ src/GRA.Controllers/ViewModel/MissionControl/Challenges/ChallengesDetailViewModel.cs
@@ -17,6 +17,7 @@ namespace GRA.Controllers.ViewModel.MissionControl.Challenges
         public IFormFile BadgeUploadImage { get; set; }
         public bool CanActivate { get; set; }
         public bool CanEditGroups { get; set; }
+        public bool CanManageEvents { get; set; }
         public bool CanViewTriggers { get; set; }
         public string MaxPointsMessage { get; set; }
         public string BadgeMakerUrl { get; set; }
@@ -24,6 +25,7 @@ namespace GRA.Controllers.ViewModel.MissionControl.Challenges
         public string BadgeMakerImage { get; set; }
         public ICollection<Trigger> DependentTriggers { get; set; }
         public List<ChallengeGroup> Groups { get; set; }
+        public List<Event> RelatedEvents { get; set; }
         public string CategoryPlaceholder { get; set; }
         public IFormFile TaskUploadFile { get; set; }
         public string TaskFilePath { get; set; }